+++ NEWS.md
@@ -11,6 +11,7 @@ New features:
 
 Bug fixes:
 + Fix false positive `PhanUnusedVariable` for variables declared before break/continue that are used after the loop. (#1985)
++ Emit `PhanUnusedVariable` for variables where definitions are shadowed by definitions in branches. (#2012)
 
 25 Sep 2018, Phan 1.0.6
 -----------------------